Move JS files from app/javascript to app/assets/javascripts
It makes sense for them to be alongside other non-Ruby assets like
stylesheets rather than up a level alongside Rails things.

Additionally, it means that the JavaScrpt linter, StandardJS[0] actually
runs.

[0] https://standardjs.com/
