+++ app/assets/javascripts/application.js
@@ -1,3 +1,3 @@
-// TODO: We don't need these imports yet.
-// require.context('govuk-frontend/dist/govuk/assets')
-// import * as GOVUKFrontend from 'govuk-frontend'
+import { initAll } from 'govuk-frontend'
+
+initAll()
