--- app/javascript/application.js
@@ -1,5 +1,0 @@
-import { initAll } from "govuk-frontend";
-import "./populate-find-ect-form"
-
-initAll();
-
